+++ recipe-guru.WebAPI/Database/ReceptKoraci.cs
@@ -3,13 +3,13 @@ namespace recipe_guru.WebAPI.Database
 {
     public partial class ReceptKoraci
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         public int RedniBrojKoraka { get; set; }
         public string Deskripcija { get; set; }
 
 
-        public long ReceptId { get; set; }
+        public int ReceptId { get; set; }
         public Recepti Recept { get; set; }
     }
 }
